filter out the rows that do have non-numbers
rationale: in some cases, the database query (for example in Cassandra) doesn't allow to
filter out the null values, etc.  In this case current version shows error.

The proposed change just throw away that rows that have non-numbers

P.S. maybe we need to log that decision, need to be discussed